Enforce exception-on-failure handling for all `Sequel::Model`s
This manually enforces the, typically default, Sequel::Model behaviour of
raising an exception when unexpected things happen in the ORM universe.

For example, failing to save an object, typecasting problems, update or delete
operations that touch zero or two rows of data, and so forth.

While these are typically the default behaviours in modern versions of Sequel,
and they represent the conservative choice -- so are likely to stay default --
that has not always been true.

Given that, it is much better to be explicit about our expectations, and
ensure we get the behaviour we desire.

Notably, we can also override this default in any model class, or model
instance, or -- most attractively -- in the individual operations.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>