Add the `constraints_validation` plugin to Razor
This is a Sequel plugin that automates some of the integration between
database and Ruby level constraints: specifically, it defines a DSL for use in
migrations that adds constraints to tables, and couples that with storing
sufficient metadata to build matching Ruby constraints at runtime.

This gives the nicer error handling of "catch early, in Ruby" without the
penalty of hand-writing duplicate tests in both the DDL and Ruby.

It also ensures that the database remains authoritative about constraints:
because all the constraints are expressed as real DML constraints, we avoid
the risk that a mistimed application of psql corrupts data.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>