Integrate `queue_classic` into razor-server
This brings in the `queue_classic` gem: a satisfactory implementation of
background task queuing using the PostgreSQL database.  While not generally
the ideal choice, QC carefully implements reasonably efficient and correct
locking and notification to reduce the cost.

This has the advantage of being predominantly in-process, and while not as
optimal as a queue that could run correctly without any service, it is only
prone to failure when *all* data-touching operations would fail.

This eliminates a robustness failure point that we would hit using, for
example, a second external process queue solution -- when we had to solve the
challenge of the database being up, but the queue down, or vice-versa.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>