Add `queue_classic` migration
This adds the database setup -- and teardown -- from QC to a migration,
allowing it to be run as part of the normal Sequel build and cleanup of
the database.

It is a distinct migration script as the QC code maintains transactionality
internally, and we want to disable Sequel migration transactions for this
script -- but only this script!

Signed-off-by: Daniel Pittman <daniel@rimspace.net>