Add courtesy support for an additional local Gemfile
Like a white courtesy phone, this makes life just a touch nicer: it allows a
local `Gemfile.local` to exist, and to be evaluated, when using bundler.

This allows local development gems such as additional rspec formatters,
debuggers, and other tools to be introduced without having to push them into
the public realm, forcing them on everyone -- and without having to constantly
hack on the main Gemfile, or maintain a second copy of it.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>