Implement the new image schema in the database
This updates the database schema, in the existing migration, to reflect the
current needs of the Image model as part of the command.  This removes a pile
of the existing data, which is not actually part of the new world
understanding of how images operate.

This also ties in the appropriate validation, ensuring that we have correct
data -- including assertions that the data passed to the command matches the
publicly modifiable parts of the model.

Some of the validation code is a little convoluted, to work around limits of
regular expressions that are compatible with both Ruby and PostgreSQL
semantics.  The great thing about POSIX regular expressions is how wonderfully
varied the implementations turn out to be.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>