My big, exciting, fabulous TorqueBox migration!
This moves from MRI to JRuby under TorqueBox as the primary runtime for
razor-server.  This move brings a large pile of features to the table, such as
real messaging and background tasks, clustering support, and threads, that
allow us to effectively deliver the uptime and capabilities that
Razor deserves.

After this change it is no longer supported to run or develop Razor with MRI.

The readme includes more detailed instructions, but the gist:

1. Install JRuby 1.7.4
2. `bundle install`
3. `torquebox deploy`
4. `torquebox run`
5. `curl http://localhost:8080/api`

Signed-off-by: Daniel Pittman <daniel@rimspace.net>