Add unit tests for Razor::API::Tag
These test ensure that Razor::API::Tag#to_hash and #to_json return the expected
values, and that the output of the two is the same given that #to_hash's keys
are symbols and JSON.parse(#to_json) returns string keys. Admittedly, the
:symbolize_names option could be passed to JSON.parse, but then the strings in
the rule array would not match since they're symbols.