Create CLI client
This is a command-line client for Razor that uses the public API to retrieve
(and eventually control) a Razor server.

Because the client interacts only through the public API, it requires only a
basic Ruby environment and the rest-client gem. This allows the client to be
used remotely as well as locally, and avoids the performance penalty
associated with loading the entire Razor codebase.

The client takes a RESTful navigation approach to object retrieval, where each
command-line argument is interpreted as a further reference to retrieve. For
example, invoking `./bin/razor tags 27` will first load the /api endpoint, then
find the object with an 'id' value 'tags'. For each further argument, it will
attempt to find an object with a matching 'id' (or 'obj_id' in the case of
reference objects) field. When the last document is retrieved, it is formatted
and printed to the terminal.