Add helper functions for client API
The Razor::View module provides Sinatra helper functions to render database
objects as hashes, which can then be returned to the client as JSON.

Each returned hash can be thought of as an extension of a basic 'object' class,
which has the following three fields:

    - 'id': a URL that uniquely identifies the object
    - 'spec': a URL that identifies the type of the object
    - 'name': a human-readable name for the object

Different types of objects may specify other properties by defining additional
key-value pairs.

This extended-map format allows clients to consume the API without knowing a
priori either the types of objects to be listed, or how to build URLs for the
listed objects.