Implement the database side "broker" entity
This reflects the instantiated version of the broker: when a user creates an
instance of the broker from the on-disk representation, they do so by giving a
name to the instance, and setting configuration parameters.

This stores the parameters, and the relationship with the on-disk object, to
allow the system to be aware of the users desired relationship.

This is not yet usable, only the container to ultimately hold the data.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>