Change interpolation of id into hostnames
The policy's hostname_pattern can now contain a '${id}' (rather than '%n')
to put the node's id into the hostname. This change is mostly there to
leave us some room to make other data available for use in
hostname_patterns - the '${expr}' syntax will just be more amenable to that
than the '%special' would be.