Add error messages with Razor::Matcher#validate
This is mostly so people don't start pulling their hair out upon finding that
something like '["=" ["fatc" "f1"] 5]' is invalid. Instead of simply finding
that validation has failed, one can now call Matcher#errors to get a list of
error messages; in this case, the message would be something like

    "uses unrecognized operator 'fatc';
    recognized operators are [...]"

The messages are meant to follow the Sequel::Model::Validations format for
the field :matcher, though it makes some of the wording awkward.