Fix bad tests in spec/app/api_spec.rb
The tests were checking, among other things, that a GET request to /api/tags
and /api/collections/policies returned JSON content, which is a valid
expectation. The problem is that both returned JSON content, though /api/tags
returned a 404 error and /api/collections/policies returned a 200. This fixes
the bad request to /api/tags and adds an additional status code check to make
sure that the test will fail if the endpoint doesn't exist.