Import `fabrication` gem for building database objects
This replaces the hand-coded object creation code in spec_helper with
Fabricator helpers.  These quickly build database objects to spec, making it
easier to get the data we need.

The primary advantage is that Fabrication provides native support for, eg,
sequences, creating multiple objects, and distinct classes of objects
being created.

This replaces the hand-coded version that was growing steadily in complexity
toward the same capabilities.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>