Create `name` alias for `hw_id` on Node objects
This introduces a special case to the `node` data type: we return the `hw_id`,
the unique API primary key, when `name` is invoked.  This is kind of a hack,
to allow the generic view code to work with nodes as well as other objects.

In the longer term we need to figure out the right model for node naming, and
replace this with that -- or improve the view support -- but for now, it works
to get the demo running.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>