Migrate API collection output to Siren
JSON Siren (https://github.com/kevinswiber/siren) is a hypermedia specification
for definiting entities in JSON. Siren provides a standard way of representing
both objects and collections alongside their associated actions. It is
essentially an extension and formalization of the current API object format, in
that it supports top-level objects and collections, and both embedding and
linking of objects while specifying the appropriate relation. Siren uses
a 'class' array in place of the current 'spec' property, which allows for
classifying something like a list with both the 'collection' class and the
class of the contained objects.

This patch replaces all objects returned by API calls with their JSON Siren
equivalents, and modifies the CLI client to accept and print Siren entities.

Note: THIS BREAKS THE CLI CLIENT ACTION/COMMAND FUNCTIONALITY. This will be
fixed with a future commit.