Add action navigation to CLI
This allows treating a Razor::CLI::Siren::Action object as a valid navigation
destination. The point is to enable actions to have help messages and such
outside of their container entity. This is a shift from "action as a part of an
entity" to "action bundled with an entity to provide context".