Use OptionParser for CLI help and command parsing
Instead of trying to reinvent the wheel for command-line option parsing, it
makes a lot more sense to use Ruby's built-in OptionParser. Since all the
expected field names are given to us, they can be used to generate command-
line switches for OptionParser, which in turn provides nicely-formatted help
text, and detects unexpected options.