Revamp how we identify nodes
The /svc/boot action that the MK uses now accepts the MACs of the network
interfaces, serial, asset tag, and uuid sent from the microkernel
(collectively called the hw_info). The bootstrap.ipxe file collects that
information.

The /svc/checkin action now expects the internal ID of the node and uses
that to identify the node.

Nodes now store the hw_info in an array; a node matches the hw_info
provided by bootstrap.ipxe if any of the values sent in matches the node's
hw_info.