Hook up execution of the broker install script
This ties the final stage, `os_complete`, common template to the broker
installation script -- by enabling the code that would fetch it and pass it to
bash(1) for evaluation.

This should enable brokers to work correctly in the final stage of operation.

It also fixes, in passing, a minor bug: booting in a KVM virtual machine would
work correctly, but booting in anything else fell through into the
error handler.

Now there are a handful more notifications, and we correctly exit the PXE
stack and have the firmware carry on through the boot process.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>