Use a more liberal hash for carrying template parameters
Using OpenStruct to provide access to the `configuration` data in broker
templates was convenient, but limited: it gave a pleasant method based syntax
for reading the content, but it didn't provide the conveniences like
enumaration methods that a real hash would.

This adds the `hashie` gem to the project, and uses that to construct a custom
Hash subclass for the purpose -- providing convenience readers, and
indifferent access where both string and symbol keys work as expected.

This avoids rebuilding the same sort of infrastructure ourselves, while
providing an awesome DWIM experience to authors of broker scripts.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>