Rewrite Puppet broker to use Puppet Labs package repositories
The original Razor Puppet broker installed Puppet from a gem.  This,
unfortunately, was kind of a rough user experience -- because the gem didn't
do much setup for users, and because it was awfully hard to excise the gem
version if you wanted to use a real package manager.

Since Puppet Labs provides, and recommends, our own package repositories for
installing the Puppet Community Edition, our broker should follow that advice.

This implements that -- rewriting the broker install script to support our own
package repositories as far as possible.

Presently that support is limited only to Linux flavours, but there is no
reason why it can't reasonably simply extend to working correctly on other
platforms in future.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>