Add a util file for migrations
Even though migrations should be self-contained, there are some things we
want to share, e.g., the regular expressions that express our notion of
what a name or URL is.

Since migrations might run individually, or several together, and constants
defined in one migration are only available if that migration is run, we
need to put these things into a separate file.