Find repo files case-insensitively
One of the fun quirks of the ESXi installer is that it uses a pure ISO9660
filesystem on the CD, rather than anything modern.  That actually supports
only upper-case characters in filenames.

Linux maps those to lowercase when mounted, but libarchive gives us the gory
truth -- and no easy way to detect that this is an ISO9660 only image.

So, because case is fun, and because many popular platforms are case
preserving but case insensitive, we treat repo files case insensitively.

This should, I think, match exact case first, then find alternate case sets,
but ultimately it makes it possible to get the expected file for ESXi.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>