Port node checkin API to return a set of messages
This updates the node checkin process: now it returns the set of queued
messages for the node, in order.  Binding nodes also queues a `reboot`, which
will be present by the time we collect the messages at the end.

This is a theoretically, but not practically, backward compatible change.
While the data structure works, the older MK client will try and execute an
empty command and consequently fail.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>