Reboot older microkernel instances to force upgrades
This introduces a "compatibility" hack: if you have a node running the older
Microkernel, it will now be instructed to reboot immediately.

In the ideal case this will result in rolling out the new server along with
the new Microkernel image, one reboot for each machine running the older code,
and then progress into the glorious new world.

In the suboptimal case this will result in rolling out the new server but
using an older (and incompatible) Microkernel agent.  That will be told to
"compatibility reboot", and will load the same old, incompatible kernel, and
will reboot again ad infinitum.

Hopefully this will alert the admin to something being not quite right but,
worst case, it leads to additional network load and a system that more or less
works as designed.

In future additional checks that would allow us to, eg, detect an older
Microkernel instance and only apply this when an outdated client would be
attractive, of course...

Signed-off-by: Daniel Pittman <daniel@rimspace.net>