Implement MK agent request versioning and compatibility
This implements a new field, `protocol-version`, into the JSON submitted by
the MK agent.  Since this is absent in the older version of the protocol we
can treat `nil` as "old version agent", and anything else as "new version
agent".

This allows us to target our actions rather than trying to be compatible with
everything all at the same time -- a better long term strategy.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>