PowerShell script to build a WinPE Razor client image
This is a PowerShell script capable of building out a Razor client WinPE
image, which we will then use to access and run the network installer.

Right now it is able to put together enough of an image to prove that we can
build WinPE, and boot it successfully, to trigger our test client script.

Notes:
 * tested only on Windows 8.0 with ADK

Requirements:
 * Install the ADK *IN THE DEFAULT LOCATION ON DISK!*
   * c:\program files (x86)\...
 * Have PowerShell.  (I have the version that shipped with Windows 8.)
 * Run the script.  Bam.

Output:
 * A WIM file containing our modifications.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>