Implement node ID lookup from hardware ID details
This implements an API for a client to request their unique node ID, if
available, from the Razor server.  This allows the Windows client -- and any
future installer burdened by the same inability to get dynamic data fed
directly into it -- to hook up with the node definition.

That, in turn, allows it to enter the normal bootstrapping process without
having to dynamically generate an entire new Windows system image every time
we boot a node.  Probably a performance win. ;)

Signed-off-by: Daniel Pittman <daniel@rimspace.net>