Add wimboot 1.0.3 to the Windows installer
This adds the wimboot 1.0.3 binary to the Windows installer.  This is needed
to be able to boot the WIM image outfitted with the Razor client memory
efficiently, and works together with iPXE.

This was downloaded from the binary release available online at the iPXE
website: http://ipxe.org/wimboot#download

This is GPLv2 software, and source code for this binary is available at the
same location.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>