Fix Kickstart boot arguments for multiple interfaces
The arguments to Kickstart to inform it of the NIC we booted from are more
complex than what we had encoded -- the BOOTIF argument is actually a
composite with two values:

  ${ARP HTYPE code}-${MAC with colons}

We were only supplying the MAC address, which meant that the installer was not
matching our boot interface correctly.  This updates that to (a) just get the
data directly from iPXE rather than our Razor server database, and (b) to pass
it with the "correct" extra information.

Technically, I just hard-coded "this is Ethernet" into the thing, which should
work fine since everything including wireless that users can reasonably boot
from to date is actually Ethernet as far as the kernel is concerned.
When this breaks for you, feel free to yell at me -- then tell me how to get
the right HTYPE code out of iPXE. ;)

[ptaswehla](https://github.com/ptaswehla) independently found and reported the
same thing, so due credit: that would have solved the problem too. :)

Signed-off-by: Daniel Pittman <daniel@rimspace.net>