Add support for Debian 7.2 multiarch netboot CD
This enhances the Debian installer to detect and (hopefully) support the new
Debian 7.2 multiarch installer CD image available as the default download link
on the debian.org homepage.

The biggest change is that this includes both an i386 and an AMD64 kernel, and
package set, on a single disk.  Optimally, this actually means that we can
detect at runtime what CPU instruction set the client supports and boot the
correct image.

This makes, I believe, this the first multi-arch capable installer.  Which the
rest of the system won't yet support, but so it goes...

This fixes issue #82, where https://github.com/eglute discovered and reported
the failure.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>