Augment Node Views
Augmenting node views.  Node summary view will show hostname and IP
address to assist with identifying a specific node in the list. The
detail view of a specific node will show the bound statis as well as the
policy assuming it has noe been removed.  If it has a message to that
effect will show in place of the policy.