/svc/boot: produce proper errors when params are bad
If we get a request without params, or one where none of the params are in
the lookup criteria configured in match_nodes_on, we need to complain
cleanly.

Before this change, either kind of request would lead to an internal server
error. Now it leads to a HTTP status code of 400 and a log entry that a bad
request was made.