Upgrade to TorqueBox 3.0.1
This brings in a number of testing fixes, as well as upgrading to JRuby 1.7.8
in the embedded environment.  The TorqueBox gem source is additionally
removed, as the issue preventing their gems hitting rubygems.org was resolved,
so we can use the single authoritative source for gems.

This also bumps the gemset engine version to 1.7.8, matching external and
internal versions of JRuby demanded by the system.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>