Exclude the `vendor` directory from coverage analysis
When we run our spec tests using a bundler installed collection of gems, they
wind up under the `vendor` directory.  This was included in coverage analysis
by default, which included a whole pile of different code that we don't care
about in our reports.

This restructures the excludes and adds the vendor directory to the filter
list, so coverage doesn't touch it any longer.  No functional changes beyond that.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>