Allow Policies to Seed Metadata on Nodes
Policies can have node_metadata set that will be added to nodes when
they bind to the policy.  The policies metadata is NON-AUTHORITIVE in
that it will only ADD NEW keys.  It will not replace or update the
values of existing keys.  The mindset here is that any exisiting
metadata would have been manually applied by someone who new what they
wanted.