Add Apache Shiro based RBAC to razor-server
This adds the JAR files for Apache Shiro, and dependencies, to the project.

It also wires up to initialize a Shiro security manager and ensure it is
available for use across the project.  This is wired to an external
configuration file, pleasantly addressed from our own configuration, and
is the recommended strategy for per-application security management.

It adds a rack middleware to bind the security context around
application execution, and enforces our requirement for successful HTTP
Basic auth over the API space.

It also updates the various API tests to ensure they appropriately
authenticate before invoking web requests, to mirror the new policy.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>