Track whether a node is installed or not
We now track the installation state of a node: once an installer finishes
and reports that by doing a GET against stage_done_url('finished'), we
remember the name of the policy and the current timestamp in the node. This
information will stick around until a node is bound to another policy.