Rename `installer` to `recipe` across the product
This updates the product to reflect the actual purpose of these entities: they
are really about applying some sort of process to the machine, rather than
being all about installing something.

This helps support the concrete use-cases that folks have asked for with this:

1. We would like to run a secure disk wipe process before we decommission a
   machine, as part of putting it back into the pool -- but only for machines
   that held certain classes of user data.

2. We would like to upgrade the firmware, or configure the hardware RAID
   controller, on a machine, then install an OS on it.

Both of these are best served by running another recipe against the machine at
some point in the life-cycle, not running an "installer" that actually doesn't
install anything.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>