API change: return objects instead of raw arrays
In the places where the API used to return raw arrays (e.g. all the
collections), now return an object where the actual array is in the 'items'
property. This allows us to annotate such collections with additional
properties in the future without breaking API backwards compatibility.

The corresponding change in razor-client is in commit 07ae04a6