Implement soft and hard IPMI reset for nodes
This adds a command that queues a reset operation to the target node, sent via
the IPMI system, and executes it as soon as possible.  This is an imperative
command, mostly expected to be used to, eg, trigger a node to reboot and run
through the PXE system.

The most likely case to trigger this is when the node will be caught be
another policy after the reboot and some additional action taken on it; for
example, after unbinding, a reboot is a reasonable way to get the new
policy applied.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>